Backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pdfplumber==0.10.2
chromadb>=0.5.0
google-generativeai>=0.8.5,<1.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
sentence-transformers>=2.2.2
transformers>=4.40.0
Pillow>=9.0.0
pdf2image>=1.17.0
pytesseract>=0.3.10
PyMuPDF>=1.24.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.0
